backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
  "docker>=7.1.0",
  "mypy>=1.17.1",
  "pip-audit>=2.6.3",
  "pre-commit>=4.2.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.1.1",
  "pytest-mock>=3.14.0",
  "testcontainers[neo4j]>=4.10.0",
  "types-python-slugify>=8.0.2.20240310"
]

[project]
name = "ion-nutri"
version = "1.1.0"
description = "Automação de Relatórios Nutrimetabolômicos da Ion Nutri"
authors = [
  {name = "Warley Vital", email = "warley.barbosa@edge.ufal.br"},
  {name = "Daniel Silva", email = "daniel.silva@edge.ufal.br"}
]
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
dependencies = [
  "fastapi[standard]>=0.115.12",
  "pydantic-settings>=2.0",
  "neo4j>=5.0.0,<6.0.0",
  "python-dotenv>=1.0.0",
  "structlog>=23.0.0",
  "uvicorn[standard]",
  "openai>=1.86.0",
  "mlflow>=3.4.0",
  "boto3>=1.34.0",
  "Jinja2>=3.1.3",
  "boto3>=1.34.0",
  "pandas",
  "slugify",
  "msoffcrypto-tool",
  "types-python-slugify>=8.0.2.20240310",
  "openpyxl>=3.1.0",
  "pdfkit",
  "markdown2",
  "tqdm"
]

[project.optional-dependencies]
dev = [
  "pytest"
]

[tool.mypy]
exclude = [
  "app/ml/.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = """
integration: marks integration tests
performance: marks performance tests that evaluate memory usage, execution time, etc.
llm: marks tests that interact with a live LLM API
neo4j: marks tests that require Neo4j database connection
"""

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade
  "B", # flake8-bugbear
  "D" # pydocstyle
]
ignore = [
  "E501" # line too long, handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-single-line = false
combine-as-imports = true

[tool.setuptools]
packages = ["app"]
